requirements.txt
pyrofork>=2.3.0
tgcrypto
requests
psutil
python-dotenv
